[refactor] Split create module into edit module
