+++ jira_offline/cli/main.py
@@ -19,7 +19,8 @@ from jira_offline.cli.params import filter_option, force_option, global_options
 from jira_offline.cli.project import cli_project_list
 from jira_offline.config import get_default_user_config_filepath
 from jira_offline.config.user_config import write_default_user_config
-from jira_offline.create import create_issue, import_csv, import_jsonlines, patch_issue_from_dict
+from jira_offline.create import create_issue, import_csv, import_jsonlines
+from jira_offline.edit import patch_issue_from_dict
 from jira_offline.exceptions import (BadProjectMetaUri, EditorFieldParseFailed, EditorNoChanges,
                                      FailedPullingProjectMeta, JiraApiError, NoInputDuringImport)
 from jira_offline.jira import jira
