+++ test/conftest.py
@@ -14,7 +14,7 @@ import requests
 from tzlocal import get_localzone
 
 from jira_offline.jira import Jira
-from jira_offline.create import get_unused_customfields
+from jira_offline.edit import get_unused_customfields
 from jira_offline.models import AppConfig, CustomFields, IssueType, ProjectMeta, Sprint
 from jira_offline.utils.cli import _get_issue, _get_project
